Began Adding Rigidbody
Have began to add in map borders, and correcting the playerMovement script from tansform.position to rigidbody2d. This is so the collision on items, objects and borders works, rather than forcing the character through the walls.
Also began adding stones to the map.
Need to add collisions to the cave entrance and the house doorway, that will open a seperate scene for both the house and the cave.
Then will begin on UI.